+++ src/js/components/parallax.js
@@ -57,7 +57,7 @@ export default {
 
         },
 
-        events: ['scroll', 'load', 'resize']
+        events: ['scroll', 'resize']
     }
 
 };