+++ src/js/components/slider.js
@@ -124,7 +124,7 @@ export default {
 
         },
 
-        events: ['load', 'resize']
+        events: ['resize']
 
     },