+++ src/js/core/core.js
@@ -5,10 +5,9 @@ export default function (UIkit) {
     ready(() => {
 
         UIkit.update();
+        on(window, 'load resize', () => UIkit.update(null, 'resize'));
+        on(document, 'loadedmetadata load', ({target}) => UIkit.update(target, 'resize'), true);
 
-        let started = 0;
-
-        on(window, 'load resize', e => UIkit.update(null, e));
         // throttle `scroll` event (Safari triggers multiple `scroll` events per frame)
         let pending;
         on(window, 'scroll', e => {
@@ -21,9 +20,10 @@ export default function (UIkit) {
 
             const {target} = e;
             UIkit.update(target.nodeType !== 1 ? document.body : target, e.type);
-        }, {passive: true, capture: true});
-        on(document, 'loadedmetadata load', ({target}) => UIkit.update(target, 'load'), true);
 
+        }, {passive: true, capture: true});
+
+        let started = 0;
         on(document, 'animationstart', ({target}) => {
             if ((css(target, 'animationName') || '').match(/^uk-.*(left|right)/)) {