+++ src/js/core/scrollspy-nav.js
@@ -94,7 +94,7 @@ export default {
 
             },
 
-            events: ['scroll', 'load', 'resize']
+            events: ['scroll', 'resize']
 
         }