+++ src/js/core/scrollspy.js
@@ -150,7 +150,7 @@ export default {
 
             },
 
-            events: ['scroll', 'load', 'resize']
+            events: ['scroll', 'resize']
 
         }