+++ src/js/mixin/slider-reactive.js
@@ -17,7 +17,7 @@ export default {
 
         },
 
-        events: ['load', 'resize']
+        events: ['resize']
 
     }